requirements.txt
blessings==1.6
click==6.7
oauthlib==2.0.1
py==1.4.32
pytest==3.0.5
python-dotenv==0.6.1
requests==2.12.4
requests-oauthlib==0.7.0
six==1.10.0
tweepy==3.5.0
